+++ src/backend/InvenTree/InvenTree/api_version.py
@@ -1,14 +1,19 @@
 """InvenTree API version information."""
 
 # InvenTree API version
-INVENTREE_API_VERSION = 460
+INVENTREE_API_VERSION = 461
 """Increment this API version number whenever there is a significant change to the API that any clients need to know about."""
 
 INVENTREE_API_TEXT = """
 
-v460 -> 2026-02-25 : https://github.com/inventree/InvenTree/pull/9928
+v461 -> 2026-03-01 : https://github.com/inventree/InvenTree/pull/9928
     - Refactor of "PartTestTemplate" models and API endpoints
     - Add "PartTest" API endpoint
+
+v460 -> 2026-02-25 : https://github.com/inventree/InvenTree/pull/11374
+    - Adds "updated_at" field to PurchaseOrder, SalesOrder and ReturnOrder API endpoints
+    - Adds "updated_before" and "updated_after" date filters to all three order list endpoints
+    - Adds "updated_at" ordering option to all three order list endpoints
 
 v459 -> 2026-02-23 : https://github.com/inventree/InvenTree/pull/11411
     - Changed PurchaseOrderLine "auto_pricing" default value from true to false